+++ src/main/java/br/com/sgp/application/core/domain/Produto.java
@@ -3,12 +3,14 @@ package br.com.sgp.application.core.domain;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class Produto {
 
     protected Long id;
-    protected Double valor;
+    protected BigDecimal valor;
     protected Boolean entregue;
     protected Boolean prontaEntrega;
     protected Boolean chegou;
